TreesAndGraphs
Changes -

1. DirectedGraph
- Added code to use DFS search.

2. Graph
- Added DFS method. Method added uses a depth first search that
utilises the natural stack in recursion to iterate through the graph.
The trick here is to pass along a list of already visited nodes and to
only go into the recursive loop again IF it hasn’t already been
visited, otherwise we will continuously visit and cause an exception.