TreesAndGraphs
Changes -

1. binarySearchTree
- Added notes as to how to create a tree with minimal height from a
sorted array. The algorithm requires to split the array into the left
and right partitions which will be used to build the left and right
sides of the tree.

2. BinarySearchTree
- Added the code to recursively create a BST from a sorted array. The
key here is to obtain the middle index correctly which is low + high /
2 as you would with merge sort.

3. BinaryTreeDepthTraversal
- Add all the elements in each level to a linked list. If we have a
binary tree with 2 levels, we’d expect 2 linked lists. Current
implementation uses a pre-order traversal to create the lists since we
visit each node level by level.