TreesAndGraphs
Changes -

1. IsBinaryTreeABinarySearchTree
- Notes on the algorithm to evaluate if a binary tree is a BST. The key
idea here is to pass the values of its children back to the parent to
evaluate. We also need values to designate if a node is a leaf, if it
violates the BST conditions and if it satisfies.

It is a good exercise in understanding how recursion works as this
algorithm uses a post-order which works in a tree in bottom up fashion
(left, right, current).

Implemented coded example on both a binary tree and BST to ensure
correct output from algorithm.