TreesAndGraphs
Changes -

1. firstCommonAncestorBST
- Added notes on traversing a binary search tree to find the first
common ancestor. The idea here is to utilise the relationships that
BST’s have with their values. The moment where the two values you want
to search for deviate (on opposing sides of the tree) then we terminate
and return the current node.

2. firstCommonAncestorBinaryTree
- Added notes on finding the common ancestor in a binary tree. Binary
trees do not have a relationship between nodes other than having the
left and right child.

A solution would be to use a post-order traversal to search the tree.
This means we search bottom up in the tree so the nodes that we have
searched are passed back to the parent.

This allows us to get the values in the left/right subtree before we
evaluate the current node. If the nodes exist in opposite ends of the
subtree, we can return the current node.

3. Code implementation for the binary trees.