TreesAndGraphs
Changes -

1. identicalBinaryTrees
- Added notes on how to resolve this problem. The current solution
involves performing 2 separate in-order traversals, the first traversal
would be to find the node in T1 that matches the value in T2. This
found node acts as the subtree that we compare.

The second half of the algorithm involves performing a modified
in-order traversal, where we only continue down the tree if both T1
(subtree) and T2 are the same. If at any point in the recursion that
they are not the same, we immediately return false as a result.

2. IdenticalBinaryTrees
- Added code to implement the algorithm as described in the notes.