Update MiscellaneousQuestions
Changes -

1. KadanesAlgorithm
- Added implementation for Kadanes algorithm that handles input for
both positive and the case of all negative numbers.

The key to handle the case for input of all negative numbers is to take
the largest value when calculating the current, whether it be the
current value in the array or the summed value with the previous. In
the case of all negatives, it will always take the current value in the
array.