Updated ObjectOrientatedDesignQuestions
Changes -

1. callCentre
- Updated the read me file to break down the question and to define all
the objects to be used in the demo along with the relationships between
each other.

2. Project Code
- Added code that represents the defined structures in the callCentre
read me. Simple example provided that removes the employees from the
call centre once a call is taken, demonstrates the passing along of
calls through different levels if particular employees are unable to
respond to the call.